requirements.txt
streamlit
docker
python-dotenv
requests
cryptography
rclone-python
